Refactor <CreateTransactionButton /> out of <TransactionView />
